requirements.txt
# Общие зависимости
python-dotenv
loguru

# Telegram-бот
python-telegram-bot==20.6

# API-сервер
fastapi==0.100.0
uvicorn==0.23.1
sqlalchemy==2.1.0
psycopg2-binary==2.9.8

# Работа с фильмами и сериалами
# imdbpy==2023.5.31

# Рекомендательная система
scikit-surprise==1.1.1
torch==2.0.1
torchvision==0.15.2
numpy==1.24.4
pandas==2.1.0
scikit-learn==1.3.1

# Тестирование
pytest==7.4.2
pytest-asyncio==0.21.0

# Дополнительные библиотеки
requests==2.31.0
redis==5.0.0
